requirements.txt
streamlit
tensorflow
numpy
pillow
streamlit_drawable_canvas
